Some configures added

+++ src/Fresh.Desktop/Pages/ProductsPage.xaml.cs
@@ -51,6 +51,10 @@ namespace Fresh.Desktop.Pages
         private async void btnDelete_Click(object sender, RoutedEventArgs e)
         {
             ProductsView products = (ProductsView)ProductsDgUi.SelectedItem;
+            if (products == null)
+            {
+                MessageBox.Show("Please select row", "Error", MessageBoxButton.OK, MessageBoxImage.Hand); return;
+            }
             ProductPage productPage = new ProductPage();
             var result = await productPage.DeleteProduct(products);
             if (result)
@@ -73,10 +77,14 @@ namespace Fresh.Desktop.Pages
         {
             ProductPage productPage = new();
             var product = (ProductsView)ProductsDgUi.SelectedItem;
+            if (product == null)
+            {
+                MessageBox.Show("Please select row", "Error", MessageBoxButton.OK, MessageBoxImage.Hand); return;
+            }
             if (await productPage.UpdateProduct(product.Id, product))
                 MessageBox.Show("Cashier successfully updated", "Success", MessageBoxButton.OK, MessageBoxImage.Information);
             else
-                MessageBox.Show("There was wrong with update cashier", "Error", MessageBoxButton.OK, MessageBoxImage.Hand);
+                MessageBox.Show("There was wrong with update product", "Error", MessageBoxButton.OK, MessageBoxImage.Hand);
         }
